Remove support for Debian and restructre code
